+++ src/SoftPosit.Net/Posits/Internal/i64_to_p8.cs
@@ -46,7 +46,7 @@ namespace System.Numerics.Posits.Internal
 
                 fracA = (fracA ^ mask);
 
-                uiA = (byte)((0x7F ^ (0x3F >> k)) | (fracA >> (k + 1)));
+                uiA = (byte)((0x7F ^ (0x3F >> k)) | (byte)(fracA >> (k + 1)));
 
                 mask = 0x1 << k; //bitNPlusOne
                 if ((mask & fracA) != 0)
